A lot of text added
I updated the license to the same as pulseaudio library. This makes it
easier to copy-paste large chunks of docummentation - which is not that
bad actually.

I added a lot of docummentation, so you don't have to refer back to the
pulseaudio library and try to guess how it translates to the perl one.

I also added a few functions in the SampleSpec that are defined in
sample.h header file. This should make SampleSpec complete.