MakeMaker: avoid installing some source files
I like to split my modules in smaller logical chunks, but in this case I
want some of them to be built as part of the main .so file. Building
them as a separate .so makes no sense. For this reason I have to name
differently than .xs - otherwise the MakeMaker with XSMULTI enabled will
treat them as sources for separate .so files.
Unfortunately when I call them .xsi (or pretty much anything that is not
recognised explicitly by MakeMaker) they will be installed longside the
.pm files. Which will only unnecessarily polute the installation.

This is a hack. In the init_PM() method of MM_Unix there are a few
blacklisted file names that will not be considered for installation.
Most of them are for swap files, which I would rather not abuse. One
notable one is that any filename with the # character anywhere in its
name will not be considered for installation. That's why I picked the
.xs# extension for my xs files that are included and should not be
installed.

Works for me (TM)